* #10: Indexing enhancements and circular dependency app startup error fix.     * Order.ts:         - Added compound index for businessId with createdAt field for faster querying of orders by business and date.     * User.ts:         - Added index for confirmed and createdAt (most used query by admin)         - Deleted unique: true statements from schema definition for email and personalId fields and         moved them to indexes definition to avoid duplicate key errors.     * Moved enum definitions for Countries and Identifications to /src/types/index.ts which was causing a circular dependency     error since functions in identification.ts files required those enums but at the same time the User.ts model required     the functions defined in identification.ts file, thus giving an error on app startup.
