+++ src/types/index.ts
@@ -1,3 +1,16 @@
+// Country & Identification enums (shared to avoid circular deps)
+export enum Countries {
+    Peru = "Peru", 
+    Chile = "Chile", 
+    Colombia = "Colombia"
+}
+
+export enum Identifications {
+    Peru = "RUC", 
+    Chile = "RUT", 
+    Colombia = "NIT", 
+}
+
 export interface UserEmailInterface {
     email: string
     name: string
