+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/index.ts","./src/server.ts","./src/config/cors.ts","./src/config/db.ts","./src/config/resend.ts","./src/config/__mocks__/resend.ts","./src/controllers/admincontroller.ts","./src/controllers/authcontroller.ts","./src/controllers/ordercontroller.ts","./src/controllers/profilecontroller.ts","./src/emails/confirmemail.ts","./src/emails/orderemail.ts","./src/emails/status/cancelledorder.ts","./src/emails/status/deliveredorder.ts","./src/emails/status/pendingorder.ts","./src/emails/status/sentorder.ts","./src/emails/status/index.ts","./src/errors/conflict-error.ts","./src/errors/custom-error.ts","./src/errors/database-connection.ts","./src/errors/forbidden-error.ts","./src/errors/not-authorized.ts","./src/errors/not-found.ts","./src/errors/request-validation.ts","./src/errors/server-error.ts","./src/middleware/auth.ts","./src/middleware/error.ts","./src/middleware/validation.ts","./src/models/order.ts","./src/models/token.ts","./src/models/user.ts","./src/routes/authrouter.ts","./src/routes/orderrouter.ts","./src/tests/setup.ts","./src/utils/auth.ts","./src/utils/jwt.ts","./src/utils/token.ts"],"version":"5.7.3"}
+{"root":["./src/index.ts","./src/server.ts","./src/config/cors.ts","./src/config/db.ts","./src/config/resend.ts","./src/config/__mocks__/resend.ts","./src/controllers/admincontroller.ts","./src/controllers/authcontroller.ts","./src/controllers/ordercontroller.ts","./src/controllers/profilecontroller.ts","./src/emails/orderemail.ts","./src/emails/admin/confirmuseremail.ts","./src/emails/admin/index.ts","./src/emails/auth/resetpasswordemail.ts","./src/emails/auth/setpasswordemail.ts","./src/emails/auth/welcomeemail.ts","./src/emails/auth/index.ts","./src/emails/status/cancelledorder.ts","./src/emails/status/deliveredorder.ts","./src/emails/status/pendingorder.ts","./src/emails/status/sentorder.ts","./src/emails/status/index.ts","./src/errors/conflict-error.ts","./src/errors/custom-error.ts","./src/errors/database-connection.ts","./src/errors/forbidden-error.ts","./src/errors/not-authorized.ts","./src/errors/not-found.ts","./src/errors/request-validation.ts","./src/errors/server-error.ts","./src/middleware/auth.ts","./src/middleware/error.ts","./src/middleware/index.ts","./src/middleware/validation.ts","./src/models/order.ts","./src/models/token.ts","./src/models/user.ts","./src/routes/authrouter.ts","./src/routes/orderrouter.ts","./src/tests/setup.ts","./src/types/index.ts","./src/utils/auth.ts","./src/utils/identifcation.ts","./src/utils/index.ts","./src/utils/jwt.ts","./src/utils/token.ts","./src/utils/validation.ts"],"version":"5.7.3"}
